backend/requirements.txt
# Google Ads API
google-ads>=25.0.0

# Anthropic Claude
anthropic>=0.39.0

# Supabase
supabase>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Email
sendgrid>=6.10.0
# or use aiosmtplib for SMTP
aiosmtplib>=3.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.0.0

# Async support
asyncio>=3.4.0
